[#30] Rename: board 타입 파일 이름 post로 바꾸고 타입 이름도 Board에서 Post로 변경
